mcp_host/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
python-dotenv
requests
websockets

